+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API changes with their rationale
 when appropriate:
 
+### v2.14.0.0
+- **all** : Upgrade of dependencies, including Kotlin to 1.9.23
+- **data4k** : Add ability to declare child data-type fields (eg. JsonNode or submap) with data() and optionalData()
+- **data4k** : [Breaking] Renamed content() to unwrap().
+
 ### v2.13.1.0
 - **all** : Upgrade of dependencies
 - **state4k** : [New module] Simple state machine modelling